requirements.txt
quixstreams
duckdb
pandas
numpy
sentence-transformers
websocket-client
nltk
dash
plotly
WordCloud
io
base64
